requirements.txt
flask
requests
gunicorn
vaderSentiment
alpaca-py
finnhub-python

#Needed for Finbert
#torch
#transformers
#Needed for Backtesting
#backtesting


#matplotlib


#Needed for XGBoost strategy
#xgboost 
#scikit-learn 
#pandas-ta
#python-dotenv
google-generativeai
yfinance
lxml
gspread 
google-auth
dotenv